feat(execute_dbt_model): add parameters to query
