+++ pipelines/rj_smtr/br_rj_riodejaneiro_brt_gps/flows.py
@@ -29,13 +29,13 @@ from pipelines.rj_smtr.tasks import (
     get_materialization_date_range,
     # get_local_dbt_client,
     # get_raw,
-    run_dbt_model,
     # save_raw_local,
     # save_treated_local,
     set_last_run_timestamp,
     # upload_logs_to_bq,
     # bq_upload,
 )
+from pipelines.utils.execute_dbt_model.tasks import run_dbt_model
 
 # from pipelines.rj_smtr.br_rj_riodejaneiro_brt_gps.tasks import (
 #     pre_treatment_br_rj_riodejaneiro_brt_gps,
