+++ pipelines/utils/execute_dbt_model/tasks.py
@@ -5,13 +5,14 @@ Tasks related to DBT flows.
 # pylint: disable=unused-argument
 
 from datetime import timedelta
-from typing import List
+from typing import Any, Dict, List, Union
 
 from dbt_client import DbtClient
 from prefect import task
 
 from pipelines.utils.execute_dbt_model.utils import (
     get_dbt_client,
+    parse_dbt_logs,
 )
 from pipelines.constants import constants
 from pipelines.utils.utils import log
@@ -43,19 +44,60 @@ def get_k8s_dbt_client(
 )
 def run_dbt_model(
     dbt_client: DbtClient,
-    dataset_id: str,
-    table_id: str,
+    dataset_id: str = None,
+    table_id: str = None,
+    model: str = None,
+    upstream: bool = None,
+    downstream: bool = None,
+    exclude: str = None,
+    flags: str = None,
+    _vars: Union[dict, List[Dict]] = None,
     sync: bool = True,
+    wait=None,  # pylint: disable=unused-argument
 ):
     """
     Run a DBT model.
     """
-    run_command = f"run --models {dataset_id}.{table_id}"
-    log(f"Running the followinf command:\n{run_command}")
-    dbt_client.cli(
+    run_command = "dbt run"
+
+    if (dataset_id and table_id) and not model:
+        model = f"{dataset_id}.{table_id}"
+
+    # Set models and upstream/downstream for dbt
+    if model:
+        run_command += " --select "
+        if upstream:
+            run_command += "+"
+        run_command += f"{model}"
+        if downstream:
+            run_command += "+"
+
+    if exclude:
+        run_command += f" --exclude {exclude}"
+
+    if _vars:
+        log(f"Received vars:\n {_vars}\n type: {type(_vars)}")
+        if isinstance(_vars, list):
+            vars_dict = {}
+            for elem in _vars:
+                log(f"Received variable {elem}. Adding to vars")
+                vars_dict.update(elem)
+            vars_str = f'"{vars_dict}"'
+            run_command += f" --vars {vars_str}"
+        else:
+            vars_str = f'"{_vars}"'
+            run_command += f" --vars {vars_str}"
+    if flags:
+        run_command += f" {flags}"
+
+    log(f"Will run the following command:\n{run_command}")
+    logs_dict = dbt_client.cli(
         run_command,
         sync=sync,
+        logs=True,
     )
+    parse_dbt_logs(logs_dict, log_queries=True)
+    return log("Finished running dbt model")
 
 
 @task(checkpoint=False)
@@ -64,3 +106,15 @@ def is_running_at_datario(current_flow_labels: List[str]) -> bool:
     Check if the current flow is running at datario agent.
     """
     return "datario" in current_flow_labels
+
+
+@task
+def is_valid_dictionary(dictionary: Dict[Any, Any]) -> Dict[str, Any]:
+    """
+    Checks whether the dictionary is in the format Dict[str, str].
+    """
+    if not isinstance(dictionary, dict):
+        raise ValueError("The dictionary is not a dictionary.")
+    if not all(isinstance(key, str) for key in dictionary.keys()):
+        raise ValueError("The dictionary keys are not strings.")
+    return dictionary
